refactor: move NoOpParser to its own file
