+++ agent/src/lib/parser/NoOpParser.ts
@@ -1,0 +1,15 @@
+import { Parser, ParsedResponse } from '.'
+
+export default class NoOpParser implements Parser {
+
+  parse () {
+    const response: ParsedResponse = {
+      parsed: false,
+      textContent: '',
+      links: [],
+      keywords: []
+    }
+    return response
+  }
+
+}
